requirements.txt
certifi==2022.9.24
charset-normalizer==2.1.1
configparser==5.3.0
contourpy==1.0.6
cycler==0.11.0
docutils==0.19
fonttools==4.38.0
idna==3.4
kiwisolver==1.4.4
matplotlib==3.6.2
numpy==1.23.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
packaging==21.3
pandas==1.5.1
Pillow==9.3.0
protobuf==3.20.1
pycocotools==2.0.6
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.6
PyYAML==6.0
requests==2.28.1
scipy==1.9.3
six==1.16.0
statistics==1.0.3.5
tensorboardX==2.5.1
torch==1.12.0+cu102
torchaudio==0.12.0+cu102
torchmetrics==0.10.3
torchvision==0.13.0+cu102
typing_extensions==4.4.0
urllib3==1.26.12
